Major change to update and process each updated section separately
- Process previous events into eventToFullTextArray (with array of
description - descArray, array of titlePrefix that contain current
event - usedInTitlePrefix, array of event names that are in current
event - eventNameArray), which is then used to fill out
updatedEventIdToTextArray. All calendar events in
updatedEventIdToTextArray are updated
- Check each section of changed calendar event to see if it's changed.
If it's changed, get event name from eventNameArray to get calendar id
to get object from eventToFullTextArray to get usedInTitlePrefix.
Iterate through and update the section for each calendar in
usedInTitlePrefix